IPL_data_Analysis/requirements.txt
kaggle
pandas
psycopg2-binary
sqlalchemy
python-dotenv
numpy
